requirements.txt
# requirements.txt
streamlit==1.31.0
openai==1.12.0
anthropic==0.8.1
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0